backend/requirements.txt
altgraph==0.17.2
asgiref==3.8.1
certifi==2024.7.4
chardet==3.0.4
distlib==0.3.8
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==24.1
platformdirs==4.2.2
python-dotenv==1.0.1
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.3
